+++ src/types/getraenk.ts
@@ -1,0 +1,7 @@
+export type Getraenk = {
+  id: string;
+  name: string;
+  price: number;
+  category: string;
+  info?: string;
+};
