+++ src/types/grouped-item.ts
@@ -1,0 +1,7 @@
+export type GroupedItem = {
+  name: string;
+  type: 'speise' | 'getraenk';
+  count: number;
+  totalPrice: number;
+  unitPrice: number;
+};
