+++ src/types/index.ts
@@ -1,0 +1,12 @@
+export * from './getraenk';
+export * from './speise';
+export * from './item';
+export * from './grouped-item';
+export * from './person';
+
+export * from './props/getraenk-zu-person-hinzufuegen-props';
+export * from './props/getraenk-details-props';
+export * from './props/person-artikel-hinzufuegen-props';
+export * from './props/person-begleichen-props';
+export * from './props/speise-details-props';
+export * from './props/speise-zu-person-hinzufuegen-props';
