+++ src/types/item.ts
@@ -1,0 +1,6 @@
+export type Item = {
+  id: string;
+  name: string;
+  price: number;
+  type: 'speise' | 'getraenk';
+};
