+++ src/types/person.ts
@@ -1,0 +1,8 @@
+import { Item } from './item';
+
+export type Person = {
+  id: string;
+  name: string;
+  totalDebt: number;
+  items: Item[];
+};
