+++ src/types/props/getraenk-details-props.ts
@@ -1,0 +1,9 @@
+import { Getraenk } from '../getraenk';
+
+export interface GetraenkDetailsProps {
+  getraenk: Getraenk;
+  visible: boolean;
+  onClose: () => void;
+  onUpdate: (updatedGetraenk: Getraenk) => void;
+  onDelete: (id: string) => void;
+}
