+++ src/types/props/getraenk-zu-person-hinzufuegen-props.ts
@@ -1,0 +1,8 @@
+import { Getraenk } from '../getraenk';
+
+export interface GetraenkZuPersonHinzufuegenProps {
+  getraenk: Getraenk;
+  visible: boolean;
+  onClose: () => void;
+  onAddToPerson: (personId: string, getraenk: Getraenk, quantity: number) => void;
+}
