+++ src/types/props/speise-details-props.ts
@@ -1,0 +1,9 @@
+import { Speise } from '../speise';
+
+export interface SpeiseDetailsProps {
+  speise: Speise;
+  visible: boolean;
+  onClose: () => void;
+  onUpdate: (updatedSpeise: Speise) => void;
+  onDelete: (id: string) => void;
+}
