+++ src/types/props/speise-zu-person-hinzufuegen-props.ts
@@ -1,0 +1,8 @@
+import { Speise } from '../speise';
+
+export interface SpeiseZuPersonHinzufuegenProps {
+  speise: Speise;
+  visible: boolean;
+  onClose: () => void;
+  onAddToPerson: (personId: string, speise: Speise, quantity: number) => void;
+}
