+++ src/types/speise.ts
@@ -1,0 +1,7 @@
+export type Speise = {
+  id: string;
+  name: string;
+  price: number;
+  category: string;
+  info?: string;
+};
